- fix cpp file not building by renaming it - added small readme info - added more required libraries + setup koin as dependency injection - completed the repo + remote + local logic - created login dialog fragment to perform the login first - added auth interceptor for putting the auth in request header; added logger interceptor too - created the homepage and displayed a list of pets; setup pagination

+++ app/src/main/java/com/example/myapplication/domain/rest/dto/AnimalsDto.kt
@@ -10,7 +10,7 @@ data class AnimalsResponseDto(
 )
 
 data class AnimalDto(
-    @SerializedName("id") val id: Int? = null,
+    @SerializedName("id") val id: Long? = null,
     @SerializedName("organization_id") val organizationId: String? = null,
     @SerializedName("url") val url: String? = null,
     @SerializedName("type") val type: String? = null,
